Update QtCocosWindow.cpp
Switching or turning off displays on a Mac moves and resizes the application window. When reinitializing the GL context, some resources which are referenced by nodes with animation are lost. When a game loop step starts , render call a deleted resource and get  a memory access error. Need stop an animation before call gameloop and restart her after.